Create README.md
A web scraper UI gradle automation project is designed to fetch data from dynamic web tables. In this case, the project fetches data related to hockey teams and Oscar-winning movies and stores it in JSON files. The project uses UI automation to interact with dynamic web pages and extract the required information. By utilizing web scraping techniques, the project can access and retrieve specific data elements from the web pages, such as team data for hockey and the top Oscar-winning movies. Once the data is collected, it is stored in JSON files for convenient access and analysis purposes